feat: Add some movement control functionality
The basic form for some movement control

It is not complete and often overshoots, but is a good base to build on

Our plow cannot handle the top speed of 2m/s or anything close. we are going to have to fix the castor wheels or go much slower than the limit.